fix(pricing): split island component into its own module for lint
Move LandingPricing out of the island entry so the file has an export
(react-refresh/only-export-components). The entry now only hydrates.
